moving immediate geometry out to its own file
